requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
pillow==10.4.0
pydantic==2.9.2